localization_simulator/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
customtkinter==5.1.3
darkdetect==0.8.0
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
numpy==1.24.3
paho-mqtt==1.6.1
Pillow==9.5.0
pyperclip==1.8.2
ratelim==0.1.6
requests==2.31.0
six==1.16.0
tkintermapview==1.29
urllib3==2.0.3
